requirements.txt
faiss-cpu
sentence-transformers
rank-bm25
fastapi
uvicorn
numpy
pandas
scikit-learn
python-docx
transformers